Move players and player_control into proper folder
update references to headers
