+++ xlive/Blam/Engine/Networking/Session/NetworkSession.h
@@ -2,7 +2,7 @@
 
 #include "..\NetworkCommon.h"
 
-#include "Blam\Engine\Players\Players.h"
+#include "Blam/Engine/game/players.h"
 #include "Blam/Engine/game/game_globals.h"
 #include "..\Transport\NetworkChannel.h"
 #include "..\Transport\NetworkObserver.h"
