+++ xlive/Blam/Engine/objects/objects.cpp
@@ -5,7 +5,7 @@
 #include "Util\Hooks\Hook.h"
 
 #include "Blam\Engine\memory\bitstream.h"
-#include "Blam\Engine\Players\Players.h"
+#include "Blam/Engine/game/players.h"
 #include "Blam\Engine\Simulation\GameInterface\SimulationGameUnits.h"
 #include "H2MOD\Modules\OnScreenDebug\OnscreenDebug.h"
 #include "H2MOD\Modules\PlayerRepresentation\PlayerRepresentation.h"
