+++ xlive/H2MOD.h
@@ -4,7 +4,7 @@
 #include "Blam/Engine/objects/object_placement.h"
 
 #include "Blam/Engine/game/game_options.h"
-#include "Blam\Engine\Players\Players.h"
+#include "Blam/Engine/game/players.h"
 #include "Blam\Common\Common.h"
 #include "H2MOD\Modules\Tweaks\Tweaks.h"
