+++ xlive/H2MOD/Modules/CustomVariantSettings/CustomVariantSettings.cpp
@@ -5,7 +5,7 @@
 #include "Blam\Engine\memory\bitstream.h"
 #include "Blam/Engine/game/game_time.h"
 #include "Blam/Engine/physics/physics_constants.h"
-#include "Blam\Engine\Players\PlayerControls.h"
+#include "Blam/Engine/game/player_control.h"
 #include "Blam\Engine\Networking\Session\NetworkSession.h"
 #include "Blam\Engine\Networking\NetworkMessageTypeCollection.h"
 #include "Blam\Engine\memory\bitstream.h"
