+++ xlive/H2MOD/Modules/KantTesting/KantTesting.cpp
@@ -11,7 +11,7 @@
 #include "Blam\Cache\TagGroups\weapon_definition.hpp"
 #include "Blam/Engine/game/game_engine.h"
 #include "Blam/Engine/game/game_globals.h"
-#include "Blam\Engine\Players\Players.h"
+#include "Blam/Engine/game/players.h"
 #include "Blam\LazyBlam\LazyBlam.hpp"
 #include "H2MOD\Engine\Engine.h"
 #include "H2MOD\Modules\Shell\Config.h"
