+++ xlive/H2MOD/Modules/SpecialEvents/Events/Halloween.cpp
@@ -11,8 +11,8 @@
 #include "Blam/Cache/TagGroups/scenario_structure_bsp_definition.hpp"
 #include "Blam/Cache/TagGroups/scenery_definition.hpp"
 #include "Blam/Engine/game/game_globals.h"
-#include "Blam/Engine/Players/Players.h"
-#include "Blam/Engine/Objects/Objects.h"
+#include "Blam/Engine/game/players.h"
+#include "Blam/Engine/objects/objects.h"
 #include "H2MOD/Modules/EventHandler/EventHandler.hpp"
 #include "H2MOD/Modules/Shell/Config.h"
 #include "H2MOD/Tags/MetaExtender.h"
